refactor: O tipo de dado do parâmetro da função formatToBRL da classe FormatUtils foi alterado de double para BigDecimal para se adequarem ao novo tipo de dado referente ao preço

+++ back-end/src/main/java/org/uorderflow/utils/FormatUtils.java
@@ -1,5 +1,6 @@
 package org.uorderflow.utils;
 
+import java.math.BigDecimal;
 import java.text.NumberFormat;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
@@ -10,7 +11,7 @@ public final class FormatUtils {
     private static final Locale PT_BR = new Locale("pt", "BR");
     private static final DateTimeFormatter DATE_TIME_FORMATTER = DateTimeFormatter.ofPattern("dd/MM/yyyy 'às' HH:mm");
 
-    public static String formatToBRL(Double price) {
+    public static String formatToBRL(BigDecimal price) {
         if (price == null) {
             return "R$ 0,00";
         }
